[Refactor] 店舗関係のコードをstore/以下に集約
