+++ src/io/input-key-processor.cpp
@@ -20,7 +20,6 @@
 #include "cmd-action/cmd-travel.h"
 #include "cmd-action/cmd-tunnel.h"
 #include "cmd-building/cmd-building.h"
-#include "cmd-building/cmd-store.h"
 #include "cmd-io/cmd-autopick.h"
 #include "cmd-io/cmd-diary.h"
 #include "cmd-io/cmd-dump.h"
@@ -81,6 +80,7 @@
 #include "player/special-defense-types.h"
 #include "status/action-setter.h"
 #include "store/home.h"
+#include "store/cmd-store.h"
 #include "store/store-util.h"
 #include "system/floor-type-definition.h"
 #include "term/screen-processor.h"
