+++ src/market/building-initializer.cpp
@@ -1,8 +1,8 @@
 ﻿#include "market/building-initializer.h"
 #include "floor/floor-town.h"
-#include "market/articles-on-sale.h"
 #include "object/object-kind.h"
 #include "object/object-kind-hook.h"
+#include "store/articles-on-sale.h"
 #include "store/store-owners.h"
 #include "store/store-util.h"
 #include "store/store.h"
