+++ src/store/cmd-store.cpp
@@ -1,5 +1,4 @@
-﻿#include "cmd-building/cmd-store.h"
-#include "cmd-io/macro-util.h"
+﻿#include "cmd-io/macro-util.h"
 #include "core/player-redraw-types.h"
 #include "core/player-update-types.h"
 #include "core/stuff-handler.h"
@@ -16,12 +15,13 @@
 #include "inventory/inventory-object.h"
 #include "inventory/inventory-slot-types.h"
 #include "io/input-key-requester.h"
-#include "io/store-key-processor.h"
 #include "main/music-definitions-table.h"
 #include "main/sound-of-music.h"
 #include "object/object-generator.h"
 #include "object/object-info.h"
+#include "store/cmd-store.h"
 #include "store/home.h"
+#include "store/store-key-processor.h"
 #include "store/store-util.h"
 #include "store/store.h"
 #include "system/floor-type-definition.h"
