+++ src/store/gold-magnification-table.h
