+++ src/store/store-key-processor.h
